load-balancer/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.6
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
setuptools==75.8.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.45.1
